+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Add response to `OAuthError`
+
 ## 0.6.4 (December 15, 2017)
 
 * Fix issue with passing request into `Procore::Response`.